Move common logic for pipeline components to pipeline module.
Depends on #84
