+++ starfish/pipeline/registration/__init__.py
@@ -1,13 +1,13 @@
-import collections
-import functools
-
+from starfish.pipeline.pipelinecomponent import PipelineComponent
 from starfish.util.argparse import FsExistsType
 from . import _base
 from . import _fourier_shift
 
 
-class Registration(object):
-    algorithm_to_class_map = dict()
+class Registration(PipelineComponent):
+    @classmethod
+    def implementing_algorithms(cls):
+        return _base.RegistrationAlgorithmBase.__subclasses__()
 
     @classmethod
     def add_to_parser(cls, subparsers):
@@ -18,19 +18,12 @@ class Registration(object):
         register_group.set_defaults(starfish_command=Registration._cli)
         registration_subparsers = register_group.add_subparsers(dest="registration_algorithm_class")
 
-        for algorithm_cls in cls.algorithm_to_class_map.values():
+        for algorithm_cls in cls.algorithm_to_class_map().values():
             group_parser = registration_subparsers.add_parser(algorithm_cls.get_algorithm_name())
             group_parser.set_defaults(registration_algorithm_class=algorithm_cls)
             algorithm_cls.add_arguments(group_parser)
 
         cls.register_group = register_group
-
-    @classmethod
-    def run(cls, algorithm_name, stack, *args, **kwargs):
-        """Runs the registration component using the algorithm name, stack, and arguments for the specific algorithm."""
-        algorithm_cls = cls.algorithm_to_class_map[algorithm_name]
-        instance = algorithm_cls(*args, **kwargs)
-        return instance.register(stack)
 
     @classmethod
     def _cli(cls, args):
@@ -50,28 +43,6 @@ class Registration(object):
         instance.register(s)
 
         s.write(args.output)
-
-    @classmethod
-    def _ensure_algorithms_setup(cls):
-        if len(cls.algorithm_to_class_map) != 0:
-            return
-
-        queue = collections.deque(_base.RegistrationAlgorithmBase.__subclasses__())
-        while len(queue) > 0:
-            algorithm_cls = queue.popleft()
-            queue.extend(algorithm_cls.__subclasses__())
-
-            cls.algorithm_to_class_map[algorithm_cls.__name__] = algorithm_cls
-
-            setattr(cls, algorithm_cls.get_algorithm_name(), Registration._trampoline_maker(algorithm_cls))
-
-    @staticmethod
-    def _trampoline_maker(algorithm_cls):
-        def trampoline(*args, **kwargs):
-            return algorithm_cls(*args, **kwargs)
-
-        functools.update_wrapper(trampoline, algorithm_cls)
-        return trampoline
 
 
 Registration._ensure_algorithms_setup()
